requirements.txt
# Core Deep Learning
torch>=2.0.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.25.0
safetensors>=0.4.0

# VLM/LLM Inference
vllm>=0.2.0
sentencepiece>=0.1.99

# Image Processing
pillow>=9.5.0
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Knowledge Graph
rdflib>=6.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
pydantic>=2.0.0
protobuf>=3.20.0
ftfy>=6.1.0
